+++ src/upytester/content/sd/lib/upyt/cmd/mapping.py
@@ -203,138 +203,3 @@ def send(obj):
     """
     global _serial_port
     _serial_port.write(json.dumps(obj).encode() + b'\r')
-
-
-def interpret_instruction(obj: dict):
-    """
-    Interpret obj as arguments for the named instruction.
-
-    ``obj`` must be of the format::
-
-        {
-            'i': <instruction name>,
-            'a': <args list>,
-            'k': <kwargs dict>,
-        }
-
-    For example::
-
-        {'i': 'func_name', 'a': [10, 'abc'], 'k': {'x': 1, 'y': 2}}
-
-    Will internally call the instruction ``func_name`` with::
-
-        func_name(10, 'abc', x=1, y=2)
-
-    Then transmit the returned value to the host (if not ``None``).
-    """
-    instruction_name = obj.get('i')
-    if instruction_name not in _instruction_map:
-        return
-
-    # Execute function
-    func = _instruction_map[instruction_name]
-    response = func(*obj.get('a', []), **obj.get('k', {}))
-
-    # Send response (if any given)
-    if response is not None:
-        send(response)
-
-
-def interpret_new_remote_instance(obj: dict):
-    """
-    Interpret obj as constructor arguments for the named remote class.
-
-    ``obj`` must be of the format::
-
-        {
-            'rc': <class name>,  # remote class
-            'a': <args list>,
-            'k': <kwargs dict>,
-        }
-
-    For example::
-
-        {'rc': 'MyRemote', 'a': [10, 'abc'], 'k': {'x': 1, 'y': 2}}
-
-    Will internally instantiate a new instance of ``MyRemote`` with::
-
-        MyRemote(10, 'abc', x=1, y=2)
-
-    Each instance will get a unique ID, and that ID will be send back to
-    the host.
-
-    Once an instance has been created, call upon it using
-    :meth:`interpret_remote_instruction`.
-    """
-    global _remote_instance_index
-
-    # Create Instance (assign unique id)
-    cls = _remote_class_map[obj.get('rc')]
-    instance = cls(*obj.get('a', []), **obj.get('k', {}))
-
-    # Save to instance map
-    instance._upyt_id = _remote_instance_index
-    _remote_instance_index += 1
-    _remote_instance_map[instance._upyt_id] = instance
-
-    # Respond with instance ID
-    send(instance._upyt_id)
-
-
-def interpret_remote_instruction(obj: dict):
-    """
-    Interpret obj as a function call to an existing remote class instance.
-
-    ``obj`` must be a :class:`dict` of the format::
-
-        {
-            'rid': <remote instance id>,  # remote id
-            'i': <method name>,
-            'a': <args list>,
-            'k': <kwargs dict>,
-        }
-
-    When instantiating new instance of class::
-
-        {'rid': 17, 'i': 'get_thing', 'a': [10, 'abc'], 'k': {'x': 1, 'y': 2}}
-
-    Will call a previously instantiated remote which was given the id ``17``::
-
-        internal_obj.get_thing(10, 'abc', x=1, y=2)
-
-    Then transmit the returned value to the host (if not ``None``).
-    """
-    # Fetch remote instance from ID
-    instance = _remote_instance_map[obj.get('rid')]
-    attr = getattr(instance, obj.get('i'))
-    if not callable(attr):
-        raise AttributeError("non-callable attributes are not supported")
-    response = attr(*obj.get('a', []), **obj.get('k', {}))
-
-    # Send response (if any given)
-    if response is not None:
-        send(response)
-
-
-def interpret(obj):
-    """
-    Perform the action defined in the given object.
-
-    :param obj: deserialized JSON object received from host
-    :type obj: :class:`dict`
-
-    Given ``obj`` must be accepted by either :meth:`interpret_instruction`
-    or :meth:`interpret_new_remote_instance`.
-    """
-    # Find referenced function
-    if not isinstance(obj, dict):
-        return
-
-    if 'rid' in obj:
-        interpret_remote_instruction(obj)
-    elif 'rc' in obj:
-        interpret_new_remote_instance(obj)
-    elif 'i' in obj:
-        interpret_instruction(obj)
-    else:
-        pass  # ignore instruction
